Added tag jdk-18+25 for changeset 8683de5e

# Conflicts:
#	src/hotspot/cpu/aarch64/sharedRuntime_aarch64.cpp
#	src/hotspot/cpu/aarch64/templateTable_aarch64.cpp
#	src/hotspot/cpu/x86/macroAssembler_x86.cpp
#	src/hotspot/cpu/x86/sharedRuntime_x86_32.cpp
#	src/hotspot/cpu/x86/sharedRuntime_x86_64.cpp
#	src/hotspot/cpu/x86/templateTable_x86.cpp
#	src/hotspot/cpu/zero/globals_zero.hpp
#	src/hotspot/share/c1/c1_LIR.hpp
#	src/hotspot/share/cds/filemap.cpp
#	src/hotspot/share/ci/ciInstanceKlass.cpp
#	src/hotspot/share/ci/ciObjectFactory.cpp
#	src/hotspot/share/ci/ciReplay.cpp
#	src/hotspot/share/classfile/javaClasses.inline.hpp
#	src/hotspot/share/gc/g1/g1FullGCCompactionPoint.cpp
#	src/hotspot/share/gc/g1/g1FullGCOopClosures.inline.hpp
#	src/hotspot/share/gc/serial/markSweep.inline.hpp
#	src/hotspot/share/gc/shared/collectedHeap.hpp
#	src/hotspot/share/gc/shared/collectedHeap.inline.hpp
#	src/hotspot/share/memory/heapInspection.cpp
#	src/hotspot/share/oops/arrayKlass.cpp
#	src/hotspot/share/oops/instanceKlass.hpp
#	src/hotspot/share/oops/objArrayKlass.cpp
#	src/hotspot/share/oops/oop.inline.hpp
#	src/hotspot/share/oops/typeArrayOop.hpp
#	src/hotspot/share/opto/callGenerator.cpp
#	src/hotspot/share/opto/cfgnode.cpp
#	src/hotspot/share/opto/macro.hpp
#	src/hotspot/share/opto/node.hpp
#	src/hotspot/share/prims/unsafe.cpp
#	src/hotspot/share/runtime/jniHandles.cpp
#	src/java.base/share/classes/jdk/internal/reflect/UnsafeFieldAccessorImpl.java
#	src/jdk.jdi/share/classes/com/sun/tools/jdi/ObjectReferenceImpl.java
#	test/hotspot/jtreg/serviceability/dcmd/vm/ShowReflectionTargetTest.java

+++ src/hotspot/share/oops/flatArrayKlass.hpp
@@ -101,7 +101,7 @@ class FlatArrayKlass : public ArrayKlass {
 
   jint max_elements() const;
 
-  int oop_size(oop obj) const;
+  size_t oop_size(oop obj) const;
 
   // Oop Allocation
   flatArrayOop allocate(int length, TRAPS);
